virt-handler: Extract VM status update logic into separate file
Move all status-related code from vm.go to new vm_status.go file for
better code organization and maintainability. This is a pure code
extraction with no functional changes.

Extracted components:
- Volume status helper functions and update methods
- Guest info, access credentials, and agent condition updates
- Migration condition calculation (live and storage)
- Paused condition handling
- Phase calculation and status update orchestration
- Memory, backup, and SELinux status updates
- Supporting helper functions and types
